Detail album screen created by 2 recycler view

+++ app/src/main/java/com/example/artist/model/SongData.java
@@ -3,6 +3,6 @@ package com.example.artist.model;
 public class SongData extends BaseModelList{
     public String title;
     public int num;
-    public double duration;
+    public int duration;
 
 }
